Remove `irrlicht_changes` directory
