+++ src/CMakeLists.txt
@@ -362,7 +362,6 @@ add_subdirectory(network)
 add_subdirectory(script)
 add_subdirectory(unittest)
 add_subdirectory(util)
-add_subdirectory(irrlicht_changes)
 add_subdirectory(server)
 
 set(common_SRCS
