+++ src/gui/guiButton.h
@@ -7,7 +7,6 @@
 #include "IrrCompileConfig.h"
 
 #include <IGUIStaticText.h>
-#include "irrlicht_changes/static_text.h"
 #include "IGUIButton.h"
 #include "IGUISpriteBank.h"
 #include "ITexture.h"
